Backend/requirements.txt
scikit-image
tensorflow
numpy
Pillow
Django
pylatex
pdflatex
folium==0.2.1
django-cors-headers
matplotlib
pandas
opencv-python
ipython
